设计二进制日志格式，并根据日志格式实现代码

+++ src/OSUtils.h
@@ -11,7 +11,7 @@ public:
 
     static int32_t getTid();
 
-    static uint32_t realTimeMillisecond();
+    static uint64_t realTimeMillisecond();
 };
 
 }// namespace qslog
